# api/openapi.yaml
openapi: 3.0.3
info:
  title: Aegis SRE Control Plane API
  version: 1.0.0
  description: Stable product API. Provider identifiers and credentials are never part of this contract.
servers:
  - url: /api/v1
paths:
...
  /playbooks:
    get:
      operationId: listPlaybooks
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Playbook references
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybookPage'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createPlaybook
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/yaml:
            schema:
              type: string
      responses:
        '201':
          description: Created playbook reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playbook'
        default:
          $ref: '#/components/responses/Problem'
  /playbooks/validate:
    post:
      operationId: validateNewPlaybook
      requestBody:
        required: true
        content:
          application/yaml:
            schema:
              type: string
      responses:
        '200':
          description: Validation result for an unsaved native Dagu YAML document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResult'
        default:
          $ref: '#/components/responses/Problem'
  /playbooks/{playbook_id}/validate:
    parameters:
      - $ref: '#/components/parameters/PlaybookID'
    post:
      operationId: validatePlaybook
      requestBody:
        required: true
        content:
          application/yaml:
            schema:
              type: string
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResult'
        default:
          $ref: '#/components/responses/Problem'
  /playbooks/{playbook_id}:
    parameters:
      - $ref: '#/components/parameters/PlaybookID'
    get:
      operationId: getPlaybook
      responses:
        '200':
          description: Playbook with its native Dagu YAML source
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playbook'
        default:
          $ref: '#/components/responses/Problem'
    put:
      operationId: updatePlaybook
      requestBody:
        required: true
        content:
          application/yaml:
            schema:
              type: string
      responses:
        '200':
          description: Updated playbook
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playbook'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deletePlaybook
      responses:
        '204':
          description: Playbook deleted
        default:
          $ref: '#/components/responses/Problem'
  /playbooks/{playbook_id}/runs:
    parameters:
      - $ref: '#/components/parameters/PlaybookID'
    get:
      operationId: listPlaybookRuns
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Playbook runs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybookRunPage'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: startPlaybookRun
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartPlaybookRunRequest'
      responses:
        '202':
          description: Run accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybookRun'
        default:
          $ref: '#/components/responses/Problem'
  /runs/{run_id}:
    parameters:
      - $ref: '#/components/parameters/RunID'
    get:
      operationId: getPlaybookRun
      responses:
        '200':
          description: Run summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybookRun'
        default:
          $ref: '#/components/responses/Problem'
...
  /runs/{run_id}/artifacts:
    parameters:
      - $ref: '#/components/parameters/RunID'
    get:
      operationId: listPlaybookArtifacts
      responses:
        '200':
          description: Run artifacts
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Artifact'
        default:
          $ref: '#/components/responses/Problem'
  /runs/{run_id}/artifacts/preview:
    parameters:
      - $ref: '#/components/parameters/RunID'
      - $ref: '#/components/parameters/ArtifactPath'
    get:
      operationId: previewPlaybookArtifact
      responses:
        '200':
          description: Text artifact preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactPreview'
        default:
          $ref: '#/components/responses/Problem'
  /runs/{run_id}/artifacts/download:
    parameters:
      - $ref: '#/components/parameters/RunID'
      - $ref: '#/components/parameters/ArtifactPath'
    get:
      operationId: downloadPlaybookArtifact
      responses:
        '200':
          description: Artifact content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        default:
          $ref: '#/components/responses/Problem'
  /runs/{run_id}/events:
    parameters:
      - $ref: '#/components/parameters/RunID'
    get:
      operationId: streamPlaybookRunEvents
      parameters:
        - name: after_sequence
          in: query
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        '200':
          description: Server-sent run events conforming to api/events.schema.json
          content:
            text/event-stream:
              schema:
                type: string
        default:
          $ref: '#/components/responses/Problem'
components:
  parameters:
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        maxLength: 512
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
    PlaybookID:
      name: playbook_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/BusinessID'
    RunID:
      name: run_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/BusinessID'
...
    ArtifactPath:
      name: path
      in: query
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 1024
  responses:
    Problem:
      description: Stable product error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    BusinessID:
      type: string
      pattern: '^[a-z][a-z0-9_]{1,15}_[A-Za-z0-9_-]{8,64}$'
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, request_id, trace_id]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          enum: [invalid_argument, unauthenticated, forbidden, not_found, conflict, capability_unavailable, provider_timeout, provider_unavailable, internal]
        detail:
          type: string
        request_id:
          type: string
        trace_id:
          type: string
        retryable:
          type: boolean
          default: false
...
    PageMetadata:
      type: object
      additionalProperties: false
      required: [has_more]
      properties:
        next_cursor:
          type: string
        has_more:
          type: boolean
...
    Playbook:
      type: object
      additionalProperties: false
      required: [id, name, description, status, source]
      properties:
        id:
          $ref: '#/components/schemas/BusinessID'
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, disabled]
        source:
          type: string
    PlaybookSummary:
      type: object
      additionalProperties: false
      required: [id, name, description, status]
      properties:
        id:
          $ref: '#/components/schemas/BusinessID'
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, disabled]
    PlaybookPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/PlaybookSummary'
    ValidationResult:
      type: object
      additionalProperties: false
      required: [valid, errors]
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [message]
            properties:
              path:
                type: string
              message:
                type: string
    StartPlaybookRunRequest:
      type: object
      additionalProperties: false
      properties:
        parameters:
          type: object
          additionalProperties:
            type: string
    PlaybookRun:
      type: object
      additionalProperties: false
      required: [id, playbook_id, status, sequence, started_at, updated_at]
      properties:
        id:
          $ref: '#/components/schemas/BusinessID'
        playbook_id:
          $ref: '#/components/schemas/BusinessID'
        status:
          type: string
          enum: [queued, running, waiting_for_input, waiting_for_approval, succeeded, failed, cancelled]
        sequence:
          type: integer
          format: int64
          minimum: 0
        started_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
        steps:
          type: array
          items:
            $ref: '#/components/schemas/PlaybookStep'
    PlaybookRunPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/PlaybookRun'
    PlaybookStep:
      type: object
      additionalProperties: false
      required: [id, name, status]
      properties:
        id: {type: string}
        name: {type: string}
        status:
          type: string
          enum: [queued, running, waiting_for_input, waiting_for_approval, succeeded, failed, cancelled]
        started_at: {type: string, format: date-time}
        ended_at: {type: string, format: date-time}
        human_task:
          type: object
          additionalProperties: true
        approval:
          type: object
          additionalProperties: true
...
    Artifact:
      type: object
      additionalProperties: false
      required: [name, path, media_type, size]
      properties:
        name: {type: string}
        path: {type: string}
        media_type: {type: string}
        size: {type: integer, format: int64, minimum: 0}
    ArtifactPreview:
      allOf:
        - $ref: '#/components/schemas/Artifact'
        - type: object
          required: [text, truncated]
          properties:
            text: {type: string}
            truncated: {type: boolean}